refactor: move File.* to IO directory
- These two files are in the Kilo::IO namespace, so it only makes sense
  that they should be moved to that folder.
- This commit also applies some .clang-format formatting rules
